Encapsulate classes in data package
All classes in the data package were changed to default access, leaving HallOfFame public, acting as entry point to the package.

Tests rearranged to correct classes.
