+++ src/main/java/com/epam/prejap/tetris/data/package-info.java
@@ -1,4 +1,6 @@
 /**
  * data package holds classes responsible for file operations: reading, writing, representing data model.
+ *
+ * <p>public class {@link com.epam.prejap.tetris.data.HallOfFame} acts as package public entry point.
  */
 package com.epam.prejap.tetris.data;
